+++ eas-20210701/core/Models/DescribeRegionsResponseBody.cs
@@ -1,0 +1,40 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.Eas20210701.Models
+{
+    public class DescribeRegionsResponseBody : TeaModel {
+        [NameInMap("Regions")]
+        [Validation(Required=false)]
+        public List<DescribeRegionsResponseBodyRegions> Regions { get; set; }
+        public class DescribeRegionsResponseBodyRegions : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>cn-shanghai</para>
+            /// </summary>
+            [NameInMap("RegionId")]
+            [Validation(Required=false)]
+            public string RegionId { get; set; }
+
+            [NameInMap("RegionName")]
+            [Validation(Required=false)]
+            public string RegionName { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>40325405-579C-4D82****</para>
+        /// </summary>
+        [NameInMap("RequestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}