+++ eas-20210701/core/Models/ListGatewayIntranetLinkedVpcResponseBody.cs
@@ -26,6 +26,10 @@ namespace AlibabaCloud.SDK.Eas20210701.Models
         [Validation(Required=false)]
         public List<ListGatewayIntranetLinkedVpcResponseBodyIntranetLinkedVpcList> IntranetLinkedVpcList { get; set; }
         public class ListGatewayIntranetLinkedVpcResponseBodyIntranetLinkedVpcList : TeaModel {
+            [NameInMap("AuthoritativeDnsEnabled")]
+            [Validation(Required=false)]
+            public bool? AuthoritativeDnsEnabled { get; set; }
+
             /// <summary>
             /// <para>The IP address.</para>
             ///